tests are more intelligent about missing files - brought up to bioperl specs
svn path=/bioperl-pedigree/trunk/; revision=12074

+++ t/Relationships.t
@@ -1,31 +1,53 @@
 # -*-Perl-*-
 
-use Test;
+use strict;
 
 BEGIN { 
-    use vars qw($NUMTESTS);
-    $NUMTESTS = 9;
+    use vars qw($NUMTESTS $error $SKIPXML) ;
+    
+    $NUMTESTS = 8;
+    $error = 0;
+    eval { require Test; };
+    if( $@ ) {
+	use lib 't';
+    }
+    use Test;
     plan tests => $NUMTESTS;
+    eval { require Tie::IxHash;
+	   require Bio::Pedigree;
+	   require Bio::Pedigree::Group;
+	   require Bio::Pedigree::PedIO;
+       };
+    if( $@ ) {
+	print STDERR "skipping tests because Tie::IxHash is not installed\n";
+	$error = 1;
+    }
+    eval { require XML::Writer };
+    if( $@ ) {
+	$SKIPXML = 1;
+    }
 }
 
+if( $error == 1 ) { exit(0); }
 END { 
+    for ( $Test::ntest..$NUMTESTS ) {
+	skip("Skipping rest of Relationship tests",1);
+    }
     unlink('test.xml');
 }
-use Bio::Pedigree::Pedigree;
-use Bio::Pedigree::Group;
-use Bio::Pedigree::PedIO;
 use Bio::Root::IO;
-
-ok(1);
 
 my $pedio = new Bio::Pedigree::PedIO(-format => 'lapis');
 my $pedigree = $pedio->read_pedigree(-pedfile => Bio::Root::IO->catfile('t','data', 'test2.lap'));
 
 ok($pedigree->calculate_all_relationships(), 12);
-my $out = new Bio::Pedigree::PedIO(-format => 'xml');
-ok($out->write_pedigree(-pedfile  => '>test.xml',
-		     -pedigree => $pedigree));
-
+unless ( $SKIPXML ) {
+    my $out = new Bio::Pedigree::PedIO(-format => 'xml');
+    ok($out->write_pedigree(-pedfile  => '>test.xml',
+			    -pedigree => $pedigree));
+} else { 
+    skip("Skipping XML output since XML writer does not exist",1);
+}
 
 my ($group1) = $pedigree->each_Group;
 my $person   = $group1->get_Person('0101');
